+++ src/php/Api/ApiFacade.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Phel\Api;
 
 use Gacela\Framework\AbstractFacade;
-use Phel\Api\Transfer\NormalizedPhelFunction;
+use Phel\Api\Transfer\PhelFunction;
 
 /**
  * @method ApiFactory getFactory()
@@ -15,12 +15,12 @@ final class ApiFacade extends AbstractFacade implements ApiFacadeInterface
     /**
      * @param list<string> $namespaces
      *
-     * @return array<string,list<NormalizedPhelFunction>>
+     * @return array<string,list<PhelFunction>>
      */
-    public function getNormalizedGroupedFunctions(array $namespaces = []): array
+    public function getGroupedFunctions(array $namespaces = []): array
     {
         return $this->getFactory()
             ->createPhelFnNormalizer()
-            ->getNormalizedGroupedFunctions($namespaces);
+            ->getGroupedFunctions($namespaces);
     }
 }