+++ tests/php/Integration/Api/ApiFacadeTest.php
@@ -14,26 +14,21 @@ use PHPUnit\Framework\TestCase;
 
 final class ApiFacadeTest extends TestCase
 {
-    public static function setUpBeforeClass(): void
-    {
-        Registry::getInstance()->clear();
-        Symbol::resetGen();
-        GlobalEnvironmentSingleton::initializeNew();
-    }
-
-    protected function setUp(): void
-    {
-        Gacela::bootstrap(__DIR__, GacelaConfig::withPhpConfigDefault());
-    }
-
     public function test_number_of_grouped_functions(): void
     {
         // This test in isolation works, but when running with all other integration tests fails,
         // because the IntegrationTest loads already the core and all internal code, which
         // it crashes when it tries to load all phel funcs again. See: `PhelFnLoader::loadAllPhelFunctions()`
         $this->markTestSkipped('Useful to debug the facade, but useless to keep it in the CI');
+
+        Gacela::bootstrap(__DIR__, GacelaConfig::withPhpConfigDefault());
+
+        Registry::getInstance()->clear();
+        Symbol::resetGen();
+        GlobalEnvironmentSingleton::initializeNew();
+
         $facade = new ApiFacade();
-        $groupedFns = $facade->getNormalizedGroupedFunctions([
+        $groupedFns = $facade->getGroupedFunctions([
             'phel\\http',
             'phel\\html',
             'phel\\test',