+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file.
 
+## Unreleased
+
+* Added `ApiFacade` which exposes 
+
 ## 0.8.0 (2023-01-16)
 
 * Allow strings on `empty?` (#492)